[#40] refactor: query를 위한 repository 이름을 support로 변경
