+++ src/main/java/com/mallang/category/query/dao/CategoryDataDao.java
@@ -1,8 +1,8 @@
 package com.mallang.category.query.dao;
 
 import com.mallang.category.domain.Category;
+import com.mallang.category.query.dao.support.CategoryQuerySupport;
 import com.mallang.category.query.data.CategoryData;
-import com.mallang.category.query.repository.CategoryQueryRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -13,10 +13,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Component
 public class CategoryDataDao {
 
-    private final CategoryQueryRepository categoryQueryRepository;
+    private final CategoryQuerySupport categoryQuerySupport;
 
     public List<CategoryData> findAllByMemberId(Long memberId) {
-        List<Category> categories = categoryQueryRepository.findAllRootByMemberId(memberId);
+        List<Category> categories = categoryQuerySupport.findAllRootByMemberId(memberId);
         return categories.stream()
                 .map(CategoryData::from)
                 .toList();
