+++ src/main/java/com/mallang/comment/query/dao/supoort/CommentQuerySupport.java
@@ -1,11 +1,11 @@
-package com.mallang.comment.query.repository;
+package com.mallang.comment.query.dao.supoort;
 
 import com.mallang.comment.domain.Comment;
 import java.util.List;
 import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentQueryRepository extends JpaRepository<Comment, Long>, CommentQueryDslRepository {
+public interface CommentQuerySupport extends JpaRepository<Comment, Long> {
 
     @EntityGraph(attributePaths = "commentWriter")
     List<Comment> findCommentsByPostId(Long postId);
