--- src/main/java/com/mallang/comment/query/repository/CommentQueryDslRepository.java
@@ -1,4 +1,0 @@
-package com.mallang.comment.query.repository;
-
-public interface CommentQueryDslRepository {
-}
