--- src/main/java/com/mallang/comment/query/repository/CommentQueryDslRepositoryImpl.java
@@ -1,4 +1,0 @@
-package com.mallang.comment.query.repository;
-
-public class CommentQueryDslRepositoryImpl implements CommentQueryDslRepository {
-}
