+++ src/main/java/com/mallang/post/query/dao/support/PostQuerySupport.java
@@ -1,12 +1,13 @@
-package com.mallang.post.query.repository;
+package com.mallang.post.query.dao.support;
 
 import com.mallang.post.domain.Post;
 import com.mallang.post.exception.NotFoundPostException;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostQueryRepository extends JpaRepository<Post, Long>, PostQueryDslRepository {
+public interface PostQuerySupport extends JpaRepository<Post, Long> {
 
     default Post getById(Long id) {
         return findById(id).orElseThrow(NotFoundPostException::new);
     }
+
 }
