--- src/main/java/com/mallang/post/query/repository/PostQueryDslRepository.java
@@ -1,10 +1,0 @@
-package com.mallang.post.query.repository;
-
-import com.mallang.post.domain.Post;
-import com.mallang.post.query.data.PostSearchCond;
-import java.util.List;
-
-public interface PostQueryDslRepository {
-
-    List<Post> search(PostSearchCond cond);
-}
